+++ FRAuth/FRAuthTests/FRAuthSwiftTests/FRAuth/FRURLProtocol/FRURLProtocolTests.swift
@@ -317,6 +317,10 @@ class FRURLProtocolTests: FRBaseTest {
         //  While API keeps returning 401; TokenManagementPolicyDelegate.evaluateTokenRefresh must only be invoked twice due to maximum retry count
         XCTAssertEqual(evaluateTokenRefreshCount, 2)
     }
+    
+    func test_07_should_manual_cleanup() {
+        self.shouldCleanup = true
+    }
 }