+++ FRAuth/FRAuthTests/TestUtil/FRTestUtils.swift
@@ -2,7 +2,7 @@
 //  FRTestUtils.swift
 //  FRAuthTests
 //
-//  Copyright (c) 2019 ForgeRock. All rights reserved.
+//  Copyright (c) 2019-2020 ForgeRock. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -26,6 +26,7 @@ class FRTestUtils: XCTest {
         FRTestNetworkStubProtocol.requestIndex = 0
         FRUser._staticUser = nil
         FRDevice._staticDevice = nil
+        Browser.currentBrowser = nil
     }
     
     @objc static func startSDK(_ config: Config) {